@@ -21,7 +21,7 @@ export default {
   data() {
     return {
       tasksList: [],
-      active: null
+      active: null,
     };
   },
   computed: {
@@ -29,27 +29,25 @@ export default {
       return this.tasksList.slice().reverse();
     },
     index(id) {
-      return this.tasksList.findIndex(item => item.id == id);
-    }
+      return this.tasksList.findIndex((item) => item.id == id);
+    },
   },
   methods: {
     changeActive(id) {
-      const index = this.tasksList.findIndex(item => item.id == id);
-
       this.active = id;
     },
 
     onDeleteTask(id) {
-      const index = this.tasksList.findIndex(item => item.id == id);
+      const index = this.tasksList.findIndex((item) => item.id == id);
       this.tasksList.splice(index, 1);
     },
     onChange(task, id) {
       const str = task.trim();
       if (str.length > 0) {
-        const index = this.tasksList.findIndex(item => item.id == id);
+        const index = this.tasksList.findIndex((item) => item.id == id);
         this.tasksList[index].cont = task;
       }
-    }
+    },
   },
   created() {
     this.tasksList = this.list;
@@ -59,14 +57,14 @@ export default {
       deep: true,
       handler() {
         this.$emit("list-change", this.tasksList);
-      }
+      },
     },
-    list: function(oldVal, newVal) {
+    list: function() {
       this.tasksList = this.list;
-    }
+    },
   },
   components: {
-    task: TasksListElement
-  }
+    task: TasksListElement,
+  },
 };
 </script>
